IoT_Data_Platform/filtering_and_storage/requirements.txt
get-docker-secret==1.0.1
paho-mqtt==1.5.0
psycopg2==2.8.5
